Rearranged file structure for the database
