+++ backend/database/models/users.py
@@ -1,0 +1,13 @@
+from sqlalchemy import Column, Integer, String, DateTime, Date
+from database.base import Base
+
+
+class User(Base):
+    __tablename__ = 'users'
+
+    id = Column(Integer, primary_key=True)
+    username = Column(String(50), unique=True, nullable=False)
+    email = Column(String(100))
+    password_hash = Column(Integer, nullable=False)
+    date_created = Column(Date)
+    last_login = Column(DateTime)
