+++ backend/database/session.py
@@ -1,8 +1,11 @@
-import os
-from dotenv import load_dotenv
 from sqlalchemy import create_engine
 from sqlalchemy.orm import sessionmaker
-from sqlalchemy.ext.declarative import declarative_base
+from sqlalchemy.orm import Session
+from typing import Annotated
+from fastapi import Depends
+import os
+from dotenv import load_dotenv
+from .base import Base
 
 # Load environment variables from .env file
 load_dotenv()
@@ -19,4 +22,14 @@ engine = create_engine(URL_DATABASE)
 
 SessionLocal = sessionmaker(autocommit = False, autoflush= False, bind = engine)
 
-Base = declarative_base()
+def get_db():
+    db = SessionLocal()
+    try:
+        yield db
+    finally:
+        db.close()
+
+db_dependency = Annotated[Session, Depends(get_db)]
+
+# Create tables
+Base.metadata.create_all(bind=engine)
